improve tests for analytics
