+++ package.json
@@ -68,6 +68,7 @@
     "jest-environment-jsdom": "30.0.5",
     "jsdom": "^28.1.0",
     "lint-staged": "^15.2.10",
+    "msw": "^2.12.14",
     "nx": "22.5.4",
     "postcss": "8.4.38",
     "prettier": "^2.8.1",
